requirements.txt
pyyaml>=6.0
requests>=2.28.0
dashscope>=1.14.0
Pillow>=9.0.0
flask>=2.3.0
flask-cors>=4.0.0